refactor: move trigger form to components
