+++ web/app/components/workflow/nodes/trigger-plugin/panel.tsx
@@ -5,7 +5,7 @@ import Split from '@/app/components/workflow/nodes/_base/components/split'
 import OutputVars, { VarItem } from '@/app/components/workflow/nodes/_base/components/output-vars'
 import type { NodePanelProps } from '@/app/components/workflow/types'
 import useConfig from './use-config'
-import TriggerForm from './trigger-form'
+import TriggerForm from './components/trigger-form'
 import StructureOutputItem from '@/app/components/workflow/nodes/_base/components/variable/object-child-tree-panel/show'
 import { Type } from '../llm/types'
